+++ common/src/main/java/dev/corgitaco/ohthetreesyoullgrow/platform/ModPlatform.java
@@ -57,6 +57,7 @@ public interface ModPlatform {
 
     enum Platform {
         FORGE,
-        FABRIC
+        FABRIC,
+        NEOFORGE
     }
 }