+++ neoforge/src/main/java/dev/corgitaco/ohthetreesyoullgrow/neoforge/platform/NeoForgeModPlatform.java
@@ -1,0 +1,86 @@
+package dev.corgitaco.ohthetreesyoullgrow.neoforge.platform;
+
+import com.google.auto.service.AutoService;
+import dev.corgitaco.ohthetreesyoullgrow.Constants;
+import dev.corgitaco.ohthetreesyoullgrow.platform.ModPlatform;
+import it.unimi.dsi.fastutil.objects.Reference2ObjectOpenHashMap;
+import net.minecraft.core.BlockPos;
+import net.minecraft.core.Registry;
+import net.minecraft.resources.ResourceKey;
+import net.minecraft.util.RandomSource;
+import net.minecraft.world.level.Level;
+import net.neoforged.api.distmarker.Dist;
+import net.neoforged.bus.api.Event;
+import net.neoforged.fml.ModList;
+import net.neoforged.fml.ModLoader;
+import net.neoforged.fml.loading.FMLEnvironment;
+import net.neoforged.fml.loading.FMLPaths;
+import net.neoforged.neoforge.common.NeoForge;
+import net.neoforged.neoforge.event.EventHooks;
+import net.neoforged.neoforge.registries.DeferredRegister;
+
+import java.nio.file.Path;
+import java.util.Map;
+import java.util.function.Supplier;
+
+@AutoService(ModPlatform.class)
+public class NeoForgeModPlatform implements ModPlatform {
+
+    @Override
+    public Path configPath() {
+        return FMLPaths.CONFIGDIR.get().resolve(Constants.MOD_ID);
+    }
+
+    @Override
+    public boolean isModLoaded(String isLoaded) {
+        return ModList.get().isLoaded(isLoaded);
+    }
+
+
+    @Override
+    public String curseForgeURL() {
+        return "";
+    }
+
+    @Override
+    public boolean isClientEnvironment() {
+        return FMLEnvironment.dist == Dist.CLIENT;
+    }
+
+    @Override
+    public Platform modPlatform() {
+        return Platform.NEOFORGE;
+    }
+
+    @Override
+    public boolean hasLoadErrors() {
+        return !ModLoader.isLoadingStateValid();
+    }
+
+    @Override
+    public boolean isDevEnvironment() {
+        return !FMLEnvironment.production;
+    }
+
+    @Override
+    public String tagNameSpace() {
+        return "forge";
+    }
+
+    @Override
+    public void addTagsUpdatedListener(TagsUpdatedEvent onTagsUpdated) {
+        NeoForge.EVENT_BUS.addListener((final net.neoforged.neoforge.event.TagsUpdatedEvent event) -> onTagsUpdated.onTagsUpdated(event.getRegistryAccess()));
+    }
+
+    @Override
+    public boolean canTreeGrowWithEvent(Level level, RandomSource source, BlockPos pos) {
+        return EventHooks.blockGrowFeature(level, source, pos, null).getResult().equals(Event.Result.DENY);
+    }
+
+    public static final Map<ResourceKey<?>, DeferredRegister> CACHED = new Reference2ObjectOpenHashMap<>();
+
+    @Override
+    public <T> Supplier<T> register(Registry<? super T> registry, String name, Supplier<T> value) {
+        return CACHED.computeIfAbsent(registry.key(), key -> DeferredRegister.create(registry.key().location(), Constants.MOD_ID)).register(name, value);
+    }
+}