+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # 1.3.0
 * Rewrite Build Script
 * Generalize Registration
+* Add TreeFromStructureNBTConfig constructor that has an empty tree decorator
 
 # 1.2.0
 * Move project to Arch Loom